requirements.txt
web3==6.11.3
eth-account==0.10.0
python-dotenv==1.0.0
apscheduler==3.10.4
anthropic==0.25.0
python-telegram-bot==20.7
pandas==2.1.3
numpy==1.26.2
ta==0.11.0
requests==2.31.0
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
